% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADF.R
\name{ADF}
\alias{ADF}
\title{ADF Test}
\usage{
ADF(my_var, df, type = "type2")
}
\arguments{
\item{my_var}{The variable to be tested, with the null hypothesis stationary (random walk) and the alternative as non-stationary.}

\item{df}{The data frame where the variable is contained, e.g. "my_dataframe".}

\item{type}{The three types are "no constant, no trend", "constant, no trend" and "constant, trend". The defaulT is "type2", but there is also "type1" and "type3".}
}
\value{
The result of the ADF test, and the first difference of the variable if it is stationary
}
\description{
Conduct an ADF test and automatically find the first difference if the variable is non-stationary.
}
\examples{
file = system.file("extdata", "hseinv.txt", package="Kenometrics")
data <- read.delim(file)
data <- ADF("linvpc", data)
data <- ADF("lpop", data, type="type1")
}
