% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_data.R
\name{get_data}
\alias{get_data}
\title{Importing Data}
\usage{
get_data(name, type = "wooldridge", delim = "", header = TRUE, wd = wd, ...)
}
\arguments{
\item{name}{The filename of the dataset to be used, e.g. "hseinv" or "NASA", without the extension (i.e. "NASA", not "NASA.xlsx").  If using FRED, type the series id.}

\item{type}{The type of data, i.e "Wooldridge", "excel" or "FRED". If using FRED then the fredr_set_key() be used prior, with the series id as the name parameter.}

\item{delim}{The delimiter used for txt files, with whitespace/tab delimiter being the default.}

\item{header}{Use FALSE if your data has no header names.}

\item{wd}{The working directory for the file, if appropriate.}

\item{...}{The additional parameters either for sourcing FRED data (start/end observations) or txt/csv files.}
}
\value{
The dataset as a dataframe
}
\description{
Import data from a variety of sources
}
\examples{
data <- get_data("hseinv", "wooldridge") #add functionality for using FRED!
my_data <- get_data("NASA", "txt", "/", FALSE, wd="C:/Users/point/Documents/R/Kenometrics/R")
}
