% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_numeric.R
\name{make_numeric}
\alias{make_numeric}
\title{Convert (Zoo) Data Frame to Numeric}
\usage{
make_numeric(df, time_var = "year")
}
\arguments{
\item{df}{The data frame to be converted (which is not numeric)}

\item{time_var}{The variable in the dataset which marks the
year of the observations. E.g. "year"}
}
\value{
The original data frame but in numeric form
}
\description{
Convert (Zoo) Data Frame to Numeric
}
\examples{
#Souring the data
file = system.file("extdata", "hseinv.txt", package="Kenometrics")
data <- read.delim(file)

#Making the data time-series
TS_data <- make_ts("year", data)

#Converting back to numeric
non_TS_data <- make_numeric(TS_data, "year")
}
