% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ts.R
\name{make_ts}
\alias{make_ts}
\title{Make Time Series}
\usage{
make_ts(time_vars, df, vars = "all")
}
\arguments{
\item{time_vars}{The variable in the dataset which marks the
year of the observations. E.g. "year". If there are separate year and month
variables, use a vector with the year variable first and the month variable
second. E.g. c("year", "month")}

\item{df}{The dataframe which needs to be made a time series data}

\item{vars}{A vector of variable names, e.g. c("pop", "inv", )}
}
\value{
The new dataset
}
\description{
Convert the entire dataset to time-series data type
}
\examples{
file = system.file("extdata", "hseinv.txt", package="Kenometrics")
data <- read.delim(file)
TS_data <- make_ts("year", data)
}
