% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data for Analysis}
\format{

}
\usage{
sample_data
}
\description{
A simple \code{tibble} that is for testing and running of examples.
}
\keyword{datasets}
