% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ts.R
\name{make_ts}
\alias{make_ts}
\title{Make Time Series}
\usage{
make_ts(time_var, df)
}
\arguments{
\item{time_var}{The variable in the dataset which marks the
year of the observations. E.g. "year"}

\item{df}{The dataframe which needs to be made a time series data}
}
\value{
The new dataset
}
\description{
Convert the entire dataset to time-series data type
}
\examples{
file = system.file("extdata", "hseinv.txt", package="Kenometrics")
data <- read.delim(file)
TS_data <- make_ts("year", data)
}
